requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
passlib
yagmail
cloudinary
python-multipart
python-dotenv
slowapi
pytest
pytest-asyncio
httpx
